Move filterActions to filters module
